# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/example/ping": {
            "get": {
                "description": "無条件に\"ping OK!\"と返します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "ping疎通確認",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/family/animals": {
            "get": {
                "description": "すべての動物をMapとして取得します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "family"
                ],
                "summary": "動物全件取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/family/animals/relations": {
            "get": {
                "description": "すべての動物をリストとして返します。自身の親、子のリストをプロパティとしてそれぞれ持ちます。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "family"
                ],
                "summary": "動物および親子関係全件取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/family/children/{parentid}": {
            "get": {
                "description": "あるノードについて、その子をすべて取得します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "family"
                ],
                "summary": "子取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "親の動物id",
                        "name": "parentid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/family/parents/{childId}": {
            "get": {
                "description": "あるノードについて、その親をすべて取得します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "family"
                ],
                "summary": "親取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "子の動物ID",
                        "name": "childId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/family/relations": {
            "get": {
                "description": "すべての親子関係をMapとして取得します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "family"
                ],
                "summary": "親子関係全件取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/picture/animal/profile/{animalId}": {
            "get": {
                "description": "特定の動物IDに紐づくプロフィール写真のバイナリデータとメタデータを取得します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "picture"
                ],
                "summary": "プロフィール写真取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "動物ID",
                        "name": "animalId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "No profile picture found"
                    }
                }
            }
        }
    }
}